added Jam3 and bstanley
