+++ runner_sim.py
@@ -9,6 +9,7 @@ from ROAR.agent_module.pure_pursuit_agent import PurePursuitAgent
 from ROAR.configurations.configuration import Configuration as AgentConfig
 from ROAR.agent_module.jAM1Agent import JAM1Agent
 from ROAR.agent_module.jAM2Agent import JAM2Agent
+from ROAR.agent_module.jAM3Agent import JAM3Agent
 
 from ROAR.agent_module.special_agents.json_waypoint_generating_agent import JSONWaypointGeneratingAgent
 from ROAR.agent_module.pid_agent import PIDAgent
@@ -27,7 +28,8 @@ def main():
 
         #agent = PurePursuitAgent(vehicle=my_vehicle, agent_settings=agent_config)
         #agent = JAM1Agent(vehicle=my_vehicle, agent_settings=agent_config)
-        agent = JAM2Agent(vehicle=my_vehicle, agent_settings=agent_config)
+        #agent = JAM2Agent(vehicle=my_vehicle, agent_settings=agent_config)
+        agent = JAM3Agent(vehicle=my_vehicle, agent_settings=agent_config)
 
 
         #carla_runner.start_game_loop(agent=agent, use_manual_control=True)#*******True for manual control, False auto
